+++ src/providers/google-analytics.js
@@ -13,7 +13,8 @@ module.exports = Provider.extend({
 
   key : 'trackingId',
 
-  options : {
+  defaults : {
+    universalClient: false,
     // Your Google Analytics Tracking ID.
     trackingId : null,
     // Whether or not to track and initial pageview when initialized.
@@ -32,7 +33,16 @@ module.exports = Provider.extend({
     doubleClick : false
   },
 
+  //
+  // Initialize
+  //
+
   initialize : function (options, ready) {
+    if (options.universalClient) this.initializeUniversal(options, ready);
+    else this.initializeClassic(options, ready);
+  },
+
+  initializeClassic: function (options, ready) {
     window._gaq = window._gaq || [];
     window._gaq.push(['_setAccount', options.trackingId]);
 
@@ -71,7 +81,53 @@ module.exports = Provider.extend({
     ready();
   },
 
+  initializeUniversal: function (options, ready) {
+
+    // GA-universal lets you set your own queue name
+    var global = this.global = 'ga';
+
+    // and needs to know about this queue name in this special object
+    // so that future plugins can also operate on the object
+    window['GoogleAnalyticsObject'] = global;
+
+    // setup the global variable
+    window[global] = window[global] || function () {
+      (window[global].q = window[global].q || []).push(arguments);
+    };
+
+    // GA also needs to know the current time (all from their snippet)
+    window[global].l = 1 * new Date();
+
+    var createOpts = {};
+
+    // Apply a bunch of optional settings.
+    if (options.domain)
+      createOpts.cookieDomain = options.domain || 'none';
+    if (type(options.siteSpeedSampleRate) === 'number')
+      createOpts.siteSpeedSampleRate = options.siteSpeedSampleRate;
+    if (options.anonymizeIp)
+      ga('set', 'anonymizeIp', true);
+
+    ga('create', options.trackingId, createOpts);
+
+    if (options.initialPageview) {
+      var path, canon = canonical();
+      if (canon) path = url.parse(canon).pathname;
+      this.pageview(path);
+    }
+
+    load('//www.google-analytics.com/analytics.js');
+
+    // Google makes a queue so it's ready immediately.
+    ready();
+  },
+
+  //
+  // Track
+  //
+
   track : function (event, properties) {
+
     properties || (properties = {});
 
     var value;
@@ -83,18 +139,46 @@ module.exports = Provider.extend({
     // so if it's not there we use `'All'` as a default. We can safely push
     // undefined if the special properties don't exist. Try using revenue
     // first, but fall back to a generic `value` as well.
-    window._gaq.push([
-      '_trackEvent',
-      properties.category || 'All',
-      event,
-      properties.label,
-      Math.round(properties.revenue) || value,
-      properties.noninteraction
-    ]);
+    if (this.options.universalClient) {
+
+      var opts = {};
+      if (properties.noninteraction)
+        opts.nonInteraction = properties.noninteraction;
+
+      window[this.global](
+         'send',
+         'event',
+         properties.category || 'All',
+         event,
+         properties.label,
+         Math.round(properties.revenue) || value,
+         opts
+      );
+
+    } else {
+
+      window._gaq.push([
+        '_trackEvent',
+        properties.category || 'All',
+        event,
+        properties.label,
+        Math.round(properties.revenue) || value,
+        properties.noninteraction
+      ]);
+    }
   },
 
+
+  //
+  // Page View
+  //
+
   pageview : function (url) {
-    window._gaq.push(['_trackPageview', url]);
+    if (this.options.universalClient) {
+      window[this.global]('send', 'pageview', url);
+    } else {
+      window._gaq.push(['_trackPageview', url]);
+    }
   }
 
 });